chore(mainapp): extract source dialog
