+++ mainapp/package.json
@@ -35,6 +35,7 @@
         "@emotion/react": "^11.11.4",
         "@emotion/styled": "^11.11.5",
         "@mui/icons-material": "^5.15.15",
+        "@mui/lab": "^5.0.0-alpha.170",
         "@mui/material": "^5.15.15",
         "@mui/x-data-grid": "^6.19.9",
         "@mui/x-tree-view": "^6.17.0",
